add Navbar.css & rename BPMNModeler
